requirements.txt
Django>=4.0.0,<=4.1.0
djangorestframework>=3.12.4,<3.13.0
psycopg2>=2.9,<2.9.4
drf-spectacular>=0.22.1,<0.24.2
pytz
pillow